chore: include core logic and integrate in plugin
