+++ packages/core/package.json
@@ -2,7 +2,7 @@
   "name": "@vite-pwa/core",
   "version": "0.12.7",
   "packageManager": "pnpm@7.9.5",
-  "description": "Zero-config PWA for Vite",
+  "description": "Vite PWA Core",
   "author": "antfu <anthonyfu117@hotmail.com>",
   "license": "MIT",
   "funding": "https://github.com/sponsors/antfu",
@@ -25,7 +25,7 @@
     "vite",
     "vite-plugin"
   ],
-  "sideEffects": "false",
+  "sideEffects": false,
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
@@ -44,7 +44,12 @@
     "build": "unbuild",
     "stub": "unbuild --stub"
   },
+  "dependencies": {
+    "workbox-build": "^6.5.4"
+  },
   "devDependencies": {
+    "@types/node": "^18.7.15",
+    "@types/workbox-build": "^5.0.1",
     "vite": "^3.1.0"
   }
 }
